Refactor a bit. Create a utils file.
